requirements.txt
numpy 
Adafruit_CircuitPython_RPLIDAR
matplotlib
flask
pyserial
rplidar-roboticia
